rename `generateInvites` to `generateSearchInvites`
