+++ src/cron/generateSearchInvites.ts
