+++ src/cron/index.ts
@@ -8,7 +8,7 @@ import exportToTinybird from './exportToTinybird';
 import cleanZombieUsers from './cleanZombieUsers';
 import cleanZombieImages from './cleanZombieImages';
 import personalizedDigest from './personalizedDigest';
-import generateInvites from './generateInvites';
+import generateSearchInvites from './generateSearchInvites';
 
 export const crons: Cron[] = [
   updateViews,
@@ -20,5 +20,5 @@ export const crons: Cron[] = [
   cleanZombieUsers,
   cleanZombieImages,
   personalizedDigest,
-  generateInvites,
+  generateSearchInvites,
 ];
